# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Roster HR Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile/me:
    get:
      operationId: getMe
      tags: [profile]
      summary: Get current user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [profile]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  # ── Departments ────────────────────────────────────────────────────────────
...
  /payroll/structures:
    get:
      operationId: listPayrollStructures
      tags: [payroll]
      summary: List payroll structures
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Payroll structures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PayrollStructure"
    post:
      operationId: createPayrollStructure
      tags: [payroll]
      summary: Create payroll structure for employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollStructureInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollStructure"
  /payroll/structures/{id}:
    patch:
      operationId: updatePayrollStructure
      tags: [payroll]
      summary: Update payroll structure
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrollStructureUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayrollStructure"
  /payroll/payslips:
    get:
      operationId: listPayslips
      tags: [payroll]
      summary: List payslips
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
        - name: month
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Payslips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payslip"
    post:
      operationId: generatePayslip
      tags: [payroll]
      summary: Generate payslip for employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayslipInput"
      responses:
        "201":
          description: Generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payslip"

  # ── Performance ────────────────────────────────────────────────────────────
  /performance/reviews:
    get:
      operationId: listPerformanceReviews
      tags: [performance]
      summary: List performance reviews
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Performance reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerformanceReview"
    post:
      operationId: createPerformanceReview
      tags: [performance]
      summary: Create performance review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerformanceReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerformanceReview"
  /performance/reviews/{id}:
    patch:
      operationId: updatePerformanceReview
      tags: [performance]
      summary: Update/acknowledge performance review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerformanceReviewUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerformanceReview"
  /performance/goals:
    get:
      operationId: listGoals
      tags: [performance]
      summary: List goals
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [performance]
      summary: Create goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /performance/goals/{id}:
    patch:
      operationId: updateGoal
      tags: [performance]
      summary: Update goal progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [performance]
      summary: Delete goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Documents ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Profile:
      type: object
      required: [id, clerkId, role, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: ["string", "null"]
        fullName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, hr_admin, manager, employee]
        employeeId:
          type: ["integer", "null"]
        createdAt:
          type: string
    ProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
        avatarUrl:
          type: ["string", "null"]

    # ── Departments ──────────────────────────────────────────────────────────
...
    PayrollStructure:
      type: object
      required: [id, employeeId, basicPay, effectiveDate, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        basicPay:
          type: number
        allowances:
          type: object
        deductions:
          type: object
        grossPay:
          type: number
        netPay:
          type: number
        effectiveDate:
          type: string
        createdAt:
          type: string
    PayrollStructureInput:
      type: object
      required: [employeeId, basicPay, effectiveDate]
      properties:
        employeeId:
          type: integer
        basicPay:
          type: number
        allowances:
          type: object
        deductions:
          type: object
        effectiveDate:
          type: string
    PayrollStructureUpdate:
      type: object
      properties:
        basicPay:
          type: number
        allowances:
          type: object
        deductions:
          type: object
        effectiveDate:
          type: string
    Payslip:
      type: object
      required: [id, employeeId, month, year, grossPay, netPay, status, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        month:
          type: integer
        year:
          type: integer
        basicPay:
          type: number
        allowances:
          type: object
        deductions:
          type: object
        grossPay:
          type: number
        netPay:
          type: number
        status:
          type: string
          enum: [draft, published]
        createdAt:
          type: string
    PayslipInput:
      type: object
      required: [employeeId, month, year]
      properties:
        employeeId:
          type: integer
        month:
          type: integer
        year:
          type: integer

    # ── Performance ───────────────────────────────────────────────────────────
    PerformanceReview:
      type: object
      required: [id, employeeId, reviewerId, rating, period, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        reviewerId:
          type: integer
        reviewerName:
          type: ["string", "null"]
        rating:
          type: integer
        period:
          type: string
        strengths:
          type: ["string", "null"]
        improvements:
          type: ["string", "null"]
        employeeComment:
          type: ["string", "null"]
        acknowledged:
          type: boolean
        createdAt:
          type: string
    PerformanceReviewInput:
      type: object
      required: [employeeId, rating, period]
      properties:
        employeeId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        period:
          type: string
        strengths:
          type: ["string", "null"]
        improvements:
          type: ["string", "null"]
    PerformanceReviewUpdate:
      type: object
      properties:
        rating:
          type: integer
        strengths:
          type: ["string", "null"]
        improvements:
          type: ["string", "null"]
        employeeComment:
          type: ["string", "null"]
        acknowledged:
          type: boolean
    Goal:
      type: object
      required: [id, employeeId, title, targetDate, progress, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        targetDate:
          type: string
        progress:
          type: integer
        createdAt:
          type: string
    GoalInput:
      type: object
      required: [employeeId, title, targetDate]
      properties:
        employeeId:
          type: integer
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        targetDate:
          type: string
        progress:
          type: integer
          default: 0
    GoalUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        targetDate:
          type: string
        progress:
          type: integer
          minimum: 0
          maximum: 100

    # ── Documents ─────────────────────────────────────────────────────────────
...